+++ twitter4j-core/src/main/java/twitter4j/SiteStreamsAdapter.java
@@ -20,61 +20,61 @@ package twitter4j;/*
  * @since Twitter4J 2.1.8
  */
 public class SiteStreamsAdapter implements SiteStreamsListener {
-    public void onStatus(int forUser, Status status) {
+    public void onStatus(long forUser, Status status) {
     }
 
-    public void onDeletionNotice(int forUser, StatusDeletionNotice statusDeletionNotice) {
+    public void onDeletionNotice(long forUser, StatusDeletionNotice statusDeletionNotice) {
     }
 
-    public void onFriendList(int forUser, int[] friendIds) {
+    public void onFriendList(long forUser, long[] friendIds) {
     }
 
-    public void onFavorite(int forUser, User source, User target, Status favoritedStatus) {
+    public void onFavorite(long forUser, User source, User target, Status favoritedStatus) {
     }
 
-    public void onUnfavorite(int forUser, User source, User target, Status unfavoritedStatus) {
+    public void onUnfavorite(long forUser, User source, User target, Status unfavoritedStatus) {
     }
 
-    public void onFollow(int forUser, User source, User followedUser) {
+    public void onFollow(long forUser, User source, User followedUser) {
     }
 
-    public void onUnfollow(int forUser, User source, User followedUser) {
+    public void onUnfollow(long forUser, User source, User followedUser) {
     }
 
-    public void onDirectMessage(int forUser, DirectMessage directMessage) {
+    public void onDirectMessage(long forUser, DirectMessage directMessage) {
     }
 
-    public void onDeletionNotice(int forUser, int directMessageId, int userId) {
+    public void onDeletionNotice(long forUser, long directMessageId, long userId) {
     }
 
-    public void onUserListMemberAddition(int forUser, User addedUser, User listOwner, UserList list) {
+    public void onUserListMemberAddition(long forUser, User addedUser, User listOwner, UserList list) {
     }
 
-    public void onUserListMemberDeletion(int forUser, User deletedUser, User listOwner, UserList list) {
+    public void onUserListMemberDeletion(long forUser, User deletedUser, User listOwner, UserList list) {
     }
 
-    public void onUserListSubscription(int forUser, User subscriber, User listOwner, UserList list) {
+    public void onUserListSubscription(long forUser, User subscriber, User listOwner, UserList list) {
     }
 
-    public void onUserListUnsubscription(int forUser, User subscriber, User listOwner, UserList list) {
+    public void onUserListUnsubscription(long forUser, User subscriber, User listOwner, UserList list) {
     }
 
-    public void onUserListCreation(int forUser, User listOwner, UserList list) {
+    public void onUserListCreation(long forUser, User listOwner, UserList list) {
     }
 
-    public void onUserListUpdate(int forUser, User listOwner, UserList list) {
+    public void onUserListUpdate(long forUser, User listOwner, UserList list) {
     }
 
-    public void onUserListDeletion(int forUser, User listOwner, UserList list) {
+    public void onUserListDeletion(long forUser, User listOwner, UserList list) {
     }
 
-    public void onUserProfileUpdate(int forUser, User updatedUser) {
+    public void onUserProfileUpdate(long forUser, User updatedUser) {
     }
 
-    public void onBlock(int forUser, User source, User blockedUser) {
+    public void onBlock(long forUser, User source, User blockedUser) {
     }
 
-    public void onUnblock(int forUser, User source, User unblockedUser) {
+    public void onUnblock(long forUser, User source, User unblockedUser) {
     }
 
     public void onException(Exception ex) {
